fix: 지원 시 대표키워드 체크 (#111)

+++ src/main/java/teamficial/teamficial_be/domain/keyword/service/TeamficialLogService.java
@@ -75,6 +75,10 @@ public class TeamficialLogService {
         if (oldHeadKeywordId != null) {
             HeadKeyword oldHeadKeyword = headKeywordService.getHeadKeywordById(oldHeadKeywordId);
 
+            if (!oldHeadKeyword.getProfile().equals(profile)) {
+                throw new GeneralException(ErrorStatus.KEYWORD_FORBIDDEN);
+            }
+
             Keyword oldHeadKeyword1 = keywordService.getKeywordByUserAndKeywordName(user,oldHeadKeyword.getKeywordName());
 
             headKeywordService.delete(oldHeadKeyword);
